core: Cast timestamp to avoid overflow
On 32-bit systems, time_t is 31 bits and overflows after ~24 days. This
should fix #38.

+++ licq/src/mainloop.cpp
@@ -96,7 +96,7 @@ long long MainLoop::Private::getMonotonicClock()
 #else
   clock_gettime(CLOCK_MONOTONIC, &ts);
 #endif
-  return ts.tv_sec * 1000 + ts.tv_nsec / 1000000;
+  return static_cast<long long>(ts.tv_sec) * 1000 + ts.tv_nsec / 1000000;
 }
 
 MainLoop::MainLoop()
